Xamarin forms, API (recommender)

+++ VisiTour.Model/Models/SpecialOffers.cs
@@ -16,7 +16,7 @@ namespace VisiTour.Model
         public int? CompanyId { get; set; }
         public int? CityFromId { get; set; }
         public int? CityToId { get; set; }
-        public decimal? Price{ get; set; }
+        public int? Price{ get; set; }
 
 
         public virtual Cities CityFrom { get; set; }
